requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.0
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.3
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
dnspython==2.7.0
email_validator==2.2.0
environs==11.0.0
fastapi==0.115.5
fastapi-cli==0.0.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
orjson==3.10.11
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.41.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.6
watchfiles==0.24.0
yarl==1.17.2
